Move Type enum
